+++ lib/l10n/app_localizations.dart
@@ -139,6 +139,66 @@ abstract class AppLocalizations {
   /// In en, this message translates to:
   /// **'Connect on open'**
   String get connectOnOpenMsg;
+
+  /// No description provided for @proxyAllRuleLabel.
+  ///
+  /// In en, this message translates to:
+  /// **'Proxy All'**
+  String get proxyAllRuleLabel;
+
+  /// No description provided for @proxyGFWRuleLabel.
+  ///
+  /// In en, this message translates to:
+  /// **'Proxy GFW'**
+  String get proxyGFWRuleLabel;
+
+  /// No description provided for @bypassCNRuleLabel.
+  ///
+  /// In en, this message translates to:
+  /// **'Bypass CN'**
+  String get bypassCNRuleLabel;
+
+  /// No description provided for @bypassAllRuleLabel.
+  ///
+  /// In en, this message translates to:
+  /// **'Bypass All'**
+  String get bypassAllRuleLabel;
+
+  /// No description provided for @goWebDashboardTip.
+  ///
+  /// In en, this message translates to:
+  /// **'Open web dashboard'**
+  String get goWebDashboardTip;
+
+  /// No description provided for @tunModeLabel.
+  ///
+  /// In en, this message translates to:
+  /// **'Tun'**
+  String get tunModeLabel;
+
+  /// No description provided for @systemModeLabel.
+  ///
+  /// In en, this message translates to:
+  /// **'System'**
+  String get systemModeLabel;
+
+  /// No description provided for @mixedModeLabel.
+  ///
+  /// In en, this message translates to:
+  /// **'Mixed'**
+  String get mixedModeLabel;
+
+  /// No description provided for @proxyModeTooltip.
+  ///
+  /// In en, this message translates to:
+  /// **'System proxy usually only supports TCP and is not accepted by all applications, but Tun can handle all traffic. Mixed enables Tun and System at the same time'**
+  String get proxyModeTooltip;
+
+  /// No description provided for @newVersionMessage.
+  ///
+  /// In en, this message translates to:
+  /// **'New available version!'**
+  String get newVersionMessage;
 }
 
 class _AppLocalizationsDelegate