+++ lib/main.dart
@@ -2,7 +2,6 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:lux/app.dart';
-import 'package:lux/const/const.dart';
 import 'package:lux/core_config.dart';
 import 'package:lux/notifier.dart';
 import 'package:lux/tr.dart';
@@ -32,17 +31,15 @@ void main(args) async {
       windowManager.show();
     });
 
-
     var isDarkMode = await readTheme() == ThemeType.dark;
-    var theme = isDarkMode ? "dark" : "light";
-    var localeModel =  LocaleModel();
+    var theme = isDarkMode ? ThemeMode.dark : ThemeMode.light;
+    var clientMode = await readClientMode();
+    var localeModel = LocaleModel();
     var defaultLocaleValue = await getLocale();
     localeModel.set(defaultLocaleValue);
-    runApp(App(theme, isDarkMode ? Color(darkBackgroundColor) : Colors.white,localeModel));
+    runApp(App(theme, localeModel, clientMode));
   } catch (e) {
     await notifier.show("$e");
     exitApp();
   }
 }
-
-