+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## What's Changed
 
+### New features
+
+* feat: [Light Client Mode](https://igoogolx.github.io/lux-docs/docs/configuration/light_client_mode)
+
+
 ### Bug fixes 🐛
 
-* fix: system proxy not working when Allow Lan is enabled
-* fix: fail to check for update since v1.35.0
+* fix: fail to upgrade form winget
 
 ### Other Changes